app: handle solo play
Now a user can play alone without storing room events on the server. Our POST
endpoints return the action back to the client, which then dispatches the
action.

This does not enable offline play -- the user still has to be able to make POST
requests to the server.

A user can also choose to play solo in a game where others could join. This is
just as solo-only mode which cannot be converted to group.

Also fix bugs related to:
- displaying $ won only if the winning buzzer was correct
- focus on the last clue on the board after clicking "next clue" (local only)
- cannot change name to ""
- skip to RevealAnswerToAll if all players in game have buzzed in and gotten
  the clue wrong