app utils name: sample from letters by frequency
Sample the starting letter by how frequently it appears in the word lists. For
example, `X.* X.*` should appear infrequently (< 1/26 of the time) because
there are only one adjective and three nouns starting with X.

Generate a probability distribution of starting letter for the adjective and
noun lists, multiply them by one another and normalize them, then sample from
that probability distribution function to get the starting letter. Then, choose
a random adjective + noun starting with that letter as before.