app components prompt: fix second buzz bug
If the player previously answered this clue and their answer was wrong, their
buzz for this clue will be -1. Don't send a timeout buzz from them again for
the next set of buzzes, and instead let the rest of the players buzz in.

Also use useTimeout instead of setTimeout where possible.