app components prompt: add separate ReadWagerableClue game state
ReadWagerableClue is a simpler state than ReadClue because only one person can
buzz. There's no need to keep optimistic buzzes or submit timeouts from the
other players. There is also no re-opening of the clue if the player is wrong.

In fact, in the future we should probably eliminate "buzzing" for the wagerable
clue entirely and move to ShowAnswerToBuzzer on timeout.