app engine: wait for all players to respond before advancing
Given the following timeline:

| server timestamp | player   | action                          |
| ---------------- | -------- | ------------------------------- |
| 0ms              | player A | chooses clue                    |
| 5ms              | player A | sees clue on their client       |
| 100ms            | player B | sees clue on their client       |
| 5005ms           | player A | times out without buzzing       |
| 5020ms           | player B | buzzes in (took 4920ms to buzz) |

When player A sees player B's buzz, they'll disregard it because A themselves
already timed out.

When player B sees player A's timeout, they'll accept it because it came in
after their buzz.

This will cause the two players to be in inconsistent states. To fix this, wait
for all buzzes or timeouts to come in before transitioning to the next state.
Do not eagerly transition when receiving any timeout.

This means that the game state can no longer advance if any player becomes
inactive (i.e. closes their tab). We will handle spectators separately in the
future and make it harder to accidentally join the game.