app components preview: join game from dialog instead of on load
Previously, any page load (by a non-bot) would add a player to the game.  This
added "fake players" to the game when someone opened the link in an app browser
(e.g. Messenger) before opening it in the browser they want to play in.

Now, the player must click "join" to affirmatively join the game. In the future
we could add "join as spectator" to this dialog as well. This keeps the game
from being polluted by random visits.